Move processor tests
